requirements/run.txt
diffpy.structure
diffpy.utils
numpy
pathlib
pycifrw
scipy
xonsh
scikit-beam
matplotlib-base
sympy
typing_extensions
pydantic
pymongo
habanero
dnspython
uvicorn
fastapi
psutil
python-multipart
google-cloud-storage
bg-mpl-stylesheets
Authlib
httpx
itsdangerous
jinja2
aiofiles
